+++ src/main/java/com/samourai/whirlpool/client/mix/MixProcess.java
@@ -198,7 +198,7 @@ public class MixProcess {
     }
 
     this.inputsHash = registerOutputMixStatusNotification.getInputsHash();
-    this.receiveDestination = postmixHandler.computeDestination();
+    this.receiveDestination = postmixHandler.computeDestinationNext();
 
     String unblindedSignedBordereau64 =
         WhirlpoolProtocol.encodeBytes(clientCryptoService.unblind(signedBordereau, blindingParams));
@@ -392,6 +392,10 @@ public class MixProcess {
 
   public UtxoDetail getReceiveUtxo() {
     return receiveUtxo;
+  }
+
+  public MixDestination getReceiveDestination() {
+    return receiveDestination;
   }
 
   private boolean throwProtocolException() throws Exception {