+++ src/main/java/com/samourai/whirlpool/client/mix/handler/AbstractPostmixHandler.java
@@ -1,6 +1,8 @@
 package com.samourai.whirlpool.client.mix.handler;
 
 import com.samourai.wallet.client.indexHandler.IIndexHandler;
+import com.samourai.whirlpool.client.utils.ClientUtils;
+import com.samourai.whirlpool.client.wallet.beans.IndexRange;
 import org.bitcoinj.core.NetworkParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,16 +20,22 @@ public abstract class AbstractPostmixHandler implements IPostmixHandler {
     this.params = params;
   }
 
-  protected abstract MixDestination computeNextDestination() throws Exception;
+  protected abstract IndexRange getIndexRange();
 
   @Override
-  public MixDestination getDestination() {
-    return destination; // may be NULL
-  }
-
-  public final MixDestination computeDestination() throws Exception {
+  public final MixDestination computeDestinationNext() throws Exception {
     // use "unconfirmed" index to avoid huge index gaps on multiple mix failures
-    this.destination = computeNextDestination();
+    int index = ClientUtils.computeNextReceiveAddressIndex(getIndexHandler(), getIndexRange());
+    this.destination = computeDestination(index);
+    if (log.isDebugEnabled()) {
+      log.debug(
+          "Mixing to "
+              + destination.getType()
+              + " -> receiveAddress="
+              + destination.getAddress()
+              + ", path="
+              + destination.getPath());
+    }
     return destination;
   }
 
@@ -43,5 +51,10 @@ public abstract class AbstractPostmixHandler implements IPostmixHandler {
   public void onRegisterOutput() {
     // confirm postmix index on REGISTER_OUTPUT success
     indexHandler.confirmUnconfirmed(destination.getIndex());
+  }
+
+  @Override
+  public IIndexHandler getIndexHandler() {
+    return indexHandler;
   }
 }