+++ src/main/java/com/samourai/whirlpool/client/mix/handler/IPostmixHandler.java
@@ -1,12 +1,15 @@
 package com.samourai.whirlpool.client.mix.handler;
 
+import com.samourai.wallet.client.indexHandler.IIndexHandler;
+
 public interface IPostmixHandler {
+  MixDestination computeDestinationNext() throws Exception;
 
-  MixDestination getDestination();
-
-  MixDestination computeDestination() throws Exception;
+  MixDestination computeDestination(int index) throws Exception;
 
   void onRegisterOutput();
 
   void onMixFail();
+
+  IIndexHandler getIndexHandler();
 }