+++ src/test/java/com/samourai/whirlpool/client/test/AbstractTest.java
@@ -382,7 +382,7 @@ public class AbstractTest {
   protected WalletStateSupplier computeWalletStateSupplier() throws Exception {
     ClientUtils.createFile(STATE_FILENAME);
     WalletStateSupplier walletStateSupplier =
-        new WalletStatePersistableSupplier(new WalletStatePersisterFile(STATE_FILENAME), null);
+        new WalletStatePersistableSupplier(new WalletStatePersisterFile(STATE_FILENAME));
     walletStateSupplier.load();
     return walletStateSupplier;
   }